requirements.txt
botoy~=0.0.13
loguru~=0.5.3
tinydb~=4.2.0
requests~=2.23.0
retrying~=1.3.3
py-cpuinfo~=7.0.0
psutil~=5.7.2
PicImageSearch~=0.6.5
pixivpy~=3.5.10
beautifulsoup4~=4.9.2
